Add option to use precomputed templates
Template processing takes a huge fraction of the overall runtime and is
all on CPU. It's not clear to me whether doing this is realistic for
production use cases (e.g. computing this on a CPU-only machine and then
moving it to a GPU machine), but at least for benchmarking focused on
GPU performance this is a useful option to have.

Note that this uses the features.pkl which is already saved by the
existing code. For more stability across numpy versions, etc. we could
consider `numpy.save`/`numpy.load`.