Fork run_docker.py to just output the command
The existing ways to run alphafold are the run_alphafold.py script that
takes a ton of explicit command line arguments or run_docker.py which
takes a few and then runs them in a docker container with a whole bunch
of separately mounted volumes that all basically point to the same
directory. It's also hard coded to use Nvidia GPUs (though that wouldn't
be too hard to change, probably).

I found this adaptation useful to just generate the command needed for
testing. In an ideal world the extra logic here would probably be
subsumed into the run_alphafold.py script with default values for
various options that are dependent on the same higher-level flags, but
I'm not sure that's worth it right now, so I'm checking in this as-is.