Script to run alphafold using precomputed features
You can cheese it by giving run_alphafold.py a bunch of empty files, but
this is cleaner. This allows computing features once on a single machine
and then copying it to wherever you want to run (or benchmark) the
model. Not only does this save expensive time doing the feature search
on GPU machines, but also avoids needing to download the alphafold dbs
entirely. All you need is the model params.