Use Miniforge instead of Miniconda
Anaconda has implemented a ToS that blocks commercial use without a paid
license.

This is for the existing Nvidia image. For the ROCm image I didn't use
conda at all because I think it is way more trouble than it's worth
inside of a docker image (I'm not fond of it outside a docker image
either, tbh).