Ignore output directory
There's a warning in the README to not put the output directory in here
because then it will get put in the docker context, but that's what
dockerignore is for. I think it's simpler to have the option to keep the
files here (potentially symlinking to another disk if necessary).