Skip prediction if there are 0 model runners
This is a bit of a hack, but with this small change, you can run with
`--num_multimer_predictions_per_model=0` and just get the features
computation. This seemed easier than duplicating this script yet again.

This actually almost works already. The only place that fails is when in
`predict_structure` it looks at the contents of `prediction_result` to
decide what confidence ranking system was used. We could change this to
a check of `model_type` instead, since that's what actually determines
it, but I figured at that point we might as well skip the whole thing
altogether.