Output recycling count more accessibly
I put this in the timings json file. This is maybe not the best place
for it, especially because everything else in there has the unit
seconds, but there wasn't an obvious better place and I didn't think we
really needed more files. It *is* relevant to the timing information and
I think being able to get this small scalar out without loading the
entire model results pkl is useful. For our purposes where we're mostly
going to be looking at the timings.json, that's a convenient place to
put it.