feat: Allow consistent seed for different models
Adds `--consistent_random_seeds` flag to run every model with the same
seed.

Right now, the random seed provided by the user is used to generate a
different seed per model and per prediction (also in a somewhat strange
way). This makes some sense if your goal is to have a reproducible run,
but still do maximum exploration of the search space, but it isn't great
for our use case for a couple of reasons:

1. It makes it harder to compare different model checkpoints head to
   head because the seed creates another variable.  We want to control
   variation in the seed ourselves, so we can compare different models
   given the same initial conditions.
2. Right now, the actual seed used for each model is dependent on how
   many total runs are being done, so increasing
   `--num_predictions_per_model` changes the seed used for the first
   model run.

I preserved the ability to request multiple predictions per model, but
this will just increment the seed for subsequent runs, not change the
seed for the earlier runs. I also preserved the current behavior when
the flag isn't passed, with the idea that someone might want to
reproduce results from prior to this change. This is honestly pretty
fiddly and I'm not sure if it's worth it. The use case of different
seeds per model, but still and overall reproducible run doesn't seem
unreasonable though, so it probably should be preserved somehow. One
other idea I had was the user explicitly specifying a base seed per
model, though that's a bit annoying.