feat: flag to clear cache between predictions
This can avoid OOMs for larger complexes. It seems pretty likely that
when running each model multiple times, one would want to clear the
cache only between models, not between predictions, but that would
require shifting around the way that the list of model runners is set up
and we don't currently need that, so I went with the easy thing for now.